+++ packages/dash-core-components/src/fragments/Dropdown.react.js
@@ -46,6 +46,7 @@ export default class Dropdown extends Component {
     render() {
         const {
             id,
+            clearable,
             multi,
             options,
             setProps,
@@ -92,6 +93,8 @@ export default class Dropdown extends Component {
                         }
                     }}
                     onInputChange={search_value => setProps({search_value})}
+                    backspaceRemoves={clearable}
+                    deleteRemoves={clearable}
                     {...omit(['setProps', 'value'], this.props)}
                 />
             </div>